+++ app/Models/EventLog.php
@@ -168,7 +168,7 @@ SQL;
     public static function fetchUnpaidEvents(
         DateTimeInterface $from,
         ?DateTimeInterface $to = null,
-        int $limit = null
+        ?int $limit = null,
     ): Collection {
         $query = self::whereNotNull('event_value_currency')
             ->where('payment_status', AdPayPaymentReportProcessor::STATUS_PAYMENT_ACCEPTED)